requirements.txt
asgiref==3.2.7
attrs==19.3.0
autobahn==20.4.2
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9
daphne==2.5.0
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
zope.interface==5.1.0
